Create README.md
Analyzing sales performance is something that is always done by a company that sells goods or services. This has many benefits, such as we can compare revenue each year, or see the trend of consumer transactions from year to year. Here, I will analyze the sales performance of an online retail store with a period of 3 years.